[Refactor] LocationHandler 위치 변경
