+++ app/src/main/java/com/and04/naturealbum/ui/NatureAlbumApp.kt
@@ -18,7 +18,6 @@ import androidx.navigation.compose.rememberNavController
 import com.and04.naturealbum.data.room.Label
 import com.and04.naturealbum.ui.album.AlbumScreen
 import com.and04.naturealbum.ui.home.HomeScreen
-import com.and04.naturealbum.ui.home.LocationHandler
 import com.and04.naturealbum.ui.labelsearch.LabelSearchScreen
 import com.and04.naturealbum.ui.savephoto.SavePhotoScreen
 import com.and04.naturealbum.ui.theme.NatureAlbumTheme
